Add support for dynamic opentracing op name
The nginx  variable value can be changed at any phase of
the request. In such cases opentracing_operation_name
directive is not taking the latest value.
opentracing_operation_name directive is executed during
request init phase, so when its points to a variable X,
and if there is any update to X after init then the change
is not reflecting to opentracing_operation_name.

Testing:
 Nginx Conf Details:
  Declared the variable X in nginx conf and initialised it
  to 'abcd'.Declared the opentracing_operation_name directive
  and pointed it to $X
 Case 1:
  Updated the X to '1234' through other nginx modules (openresty)
  Jaeger UI shows operation name for the request is 1234
 Case 2:
  No change to value of X Jaeger UI shows operation name for
  the request is abcd