Add upstream address and name tags
When a request is redirected to an upstream, then it is not visible
where it went. The ngx_http_upstream module maintains a variable
"upstream_addr" for the resolved IP address + port information. It
also sets the name of the used upstream config entry to the request
as request->upstream->upstream->host.
So add that information to the new tags "upstream.address" and
"upstream.name".