Avoid logging errors when expanding a variable with a null context (resolves #124)
When opentracing is set to off in the nginx configuration and if an
opentracing variable was used, get_opentracing_context would return
null during its expansion and an error was logged. Instead I think
a NGX_ERROR should be returned so the variable is expanded to '-'.

Also I made sure the full variable name is logged instead of only
the prefix (`opentracing_context_x_datadog_trace_id` instead of
`opentracing_context_`).